Switch over Pinboard retrieval method
